修复搜索结果无法点击的bug

+++ app/src/main/java/top/orange233/memorange/bean/MemoBean.java
@@ -24,7 +24,7 @@ public class MemoBean extends LitePalSupport {
         this.content = content;
     }
 
-    public long getNumber() {
+    public int getNumber() {
         return number;
     }
